refac: Remove "Internal" namespace (#258)
* Remove Internal namespace in SqlStatement
* Remove Internal namespace for JobsApi
* Bump version
