+++ source/Databricks/documents/release-notes/release-notes.md
@@ -1,5 +1,9 @@
 # Databricks Release Notes
 
+## Version 7.1.2
+
+- The "Internal" namespace is removed, and related code is grouped in relevant namespaces.
+
 ## Version 7.1.1
 
 - No functional change.
