+++ source/Databricks/source/Jobs.UnitTests/Client/JobsApiClientTests.cs
@@ -12,13 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.Jobs.Configuration;
-using Energinet.DataHub.Core.Databricks.Jobs.Internal;
+using Energinet.DataHub.Core.Databricks.Jobs.Client;
 using FluentAssertions;
 using Moq;
 using Xunit;
 
-namespace Energinet.DataHub.Core.Databricks.Jobs.UnitTests.Internal;
+namespace Energinet.DataHub.Core.Databricks.Jobs.UnitTests.Client;
 
 public class JobsApiClientTests
 {
