+++ source/Databricks/source/SqlStatementExecution.IntegrationTests/Client/DatabricksSqlStatementClientTests.cs
@@ -12,15 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Client;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.IntegrationTests.Fixtures;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.ResponseParsers;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.IntegrationTests;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.IntegrationTests.Client;
 
 /// <summary>
 /// We use an IClassFixture to control the life cycle of the DatabricksSqlStatementApiFixture so:
