+++ source/Databricks/source/SqlStatementExecution.IntegrationTests/Fixtures/DatabricksSchemaManager.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Configuration;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Exceptions;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
