+++ source/Databricks/source/SqlStatementExecution.UnitTests/Extensions/DependencyInjection/DatabricksSqlStatementExecutionExtensionsTests.cs
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Client;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Constants;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Extensions.DependencyInjection;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Constants;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests.Helpers;
 using FluentAssertions.Execution;
 using Microsoft.Extensions.Configuration;
